+++ Plutarch/Internal/PlutusType.hs
@@ -28,7 +28,7 @@ import Generics.SOP (
   unSOP,
  )
 import Plutarch.DataRepr.Internal.Generic (MkSum (mkSum))
-import Plutarch.DataRepr.Internal.HList (IndexList)
+import Plutarch.DataRepr.Internal.HList.Utils (IndexList)
 import Plutarch.Internal (PType, S, Term, pforce, plam', punsafeCoerce, (:-->))
 import qualified Plutarch.Internal as PI
 import Plutarch.Internal.PLam ((#))