+++ plutarch-test/src/Plutarch/ApiSpec.hs
@@ -143,7 +143,7 @@ getMint =
 getCredentials :: Term s PScriptContext -> Term s (PBuiltinList PData)
 getCredentials ctx =
   let inp = pfield @"inputs" #$ pfield @"txInfo" # ctx
-   in pmap # inputCredentialHash # pfromData inp
+   in pmap # inputCredentialHash # inp
 
 {- |
   Get the hash of the Credential in an input, treating
@@ -168,7 +168,7 @@ getSym =
 checkSignatory :: Term s (PPubKeyHash :--> PScriptContext :--> PUnit)
 checkSignatory = plam $ \ph ctx' ->
   pletFields @["txInfo", "purpose"] ctx' $ \ctx -> P.do
-    PSpending _ <- pmatch . pfromData $ ctx.purpose
+    PSpending _ <- pmatch $ ctx.purpose
     let signatories = pfield @"signatories" # ctx.txInfo
     pif
       (pelem # pdata ph # pfromData signatories)
@@ -178,15 +178,16 @@ checkSignatory = plam $ \ph ctx' ->
       perror
 
 -- | `checkSignatory` implemented using `runCont`
-checkSignatoryCont :: Term s (PPubKeyHash :--> PScriptContext :--> PUnit)
+checkSignatoryCont :: forall s. Term s (PPubKeyHash :--> PScriptContext :--> PUnit)
 checkSignatoryCont = plam $ \ph ctx' ->
   pletFields @["txInfo", "purpose"] ctx' $ \ctx -> (`runCont` id) $ do
-    purpose <- cont (pmatch . pfromData $ ctx.purpose)
+    purpose <- cont (pmatch $ ctx.purpose)
     pure $ case purpose of
       PSpending _ ->
-        let signatories = pfield @"signatories" # ctx.txInfo
+        let signatories :: Term s (PBuiltinList (PAsData PPubKeyHash))
+            signatories = pfield @"signatories" # ctx.txInfo
          in pif
-              (pelem # pdata ph # pfromData signatories)
+              (pelem # pdata ph # signatories)
               -- Success!
               (pconstant ())
               -- Signature not present.
@@ -198,7 +199,7 @@ checkSignatoryCont = plam $ \ph ctx' ->
 checkSignatoryTermCont :: Term s (PPubKeyHash :--> PScriptContext :--> PUnit)
 checkSignatoryTermCont = plam $ \ph ctx' -> unTermCont $ do
   ctx <- tcont $ pletFields @["txInfo", "purpose"] ctx'
-  PSpending _ <- tcont (pmatch . pfromData $ ctx.purpose)
+  PSpending _ <- tcont (pmatch $ ctx.purpose)
   let signatories = pfield @"signatories" # ctx.txInfo
   pure $
     pif